Refactored lobby service into partial
