+++ Core/TransportManager.cs
@@ -11,7 +11,7 @@ public partial class TransportManager : Node
     {
         Instance = this;
 
-        _transportService = new SteamTransportService();
+        _transportService = new TransportService();
         SetProcess(false);
     }
